flask-app/requirements.txt
psycopg2-binary
flask-wtf
flask
redis
requests